+++ frontend_tests/node_tests/stream_events.js
@@ -25,6 +25,7 @@ set_global("stream_list", {});
 set_global("stream_muting", {});
 set_global("subs", {});
 
+const peer_data = zrequire("peer_data");
 const people = zrequire("people");
 zrequire("stream_data");
 zrequire("stream_events");
@@ -202,7 +203,6 @@ run_test("update_property", (override) => {
 
 run_test("marked_subscribed", (override) => {
     override("stream_data.subscribe_myself", noop);
-    override("stream_data.set_subscribers", noop);
     override("stream_data.update_calculated_fields", noop);
 
     // Test undefined error
@@ -302,14 +302,9 @@ run_test("marked_subscribed", (override) => {
         override("message_util.do_unread_count_updates", noop);
         override("stream_list.add_sidebar_row", noop);
 
-        with_stub((stub) => {
-            override("stream_data.set_subscribers", stub.f);
-            const user_ids = [15, 20, 25];
-            stream_events.mark_subscribed(frontend, user_ids, "");
-            const args = stub.get_args("sub", "subscribers");
-            assert.deepEqual(frontend, args.sub);
-            assert.deepEqual(user_ids, args.subscribers);
-        });
+        const user_ids = [15, 20, 25];
+        stream_events.mark_subscribed(frontend, user_ids, "");
+        assert.deepEqual(new Set(peer_data.get_subscribers(frontend.stream_id)), new Set(user_ids));
 
         // assign self as well
         with_stub((stub) => {
@@ -421,9 +416,12 @@ run_test("remove_deactivated_user_from_all_streams", () => {
 
     dev_help.can_access_subscribers = true;
 
+    // assert starting state
+    assert(!stream_data.is_user_subscribed(dev_help.stream_id, george.user_id));
+
     // verify that deactivating user should unsubscribe user from all streams
-    assert(stream_data.add_subscriber(dev_help.stream_id, george.user_id));
-    assert(dev_help.subscribers.has(george.user_id));
+    assert(peer_data.add_subscriber(dev_help.stream_id, george.user_id));
+    assert(stream_data.is_user_subscribed(dev_help.stream_id, george.user_id));
 
     stream_events.remove_deactivated_user_from_all_streams(george.user_id);