+++ zerver/middleware.py
@@ -371,6 +371,7 @@ def csrf_failure(request: HttpRequest, reason: str="") -> HttpResponse:
 
 class LocaleMiddleware(DjangoLocaleMiddleware):
     def process_response(self, request: HttpRequest, response: HttpResponse) -> HttpResponse:
+
         # This is the same as the default LocaleMiddleware, minus the
         # logic that redirects 404's that lack a prefixed language in
         # the path into having a language.  See
@@ -382,6 +383,13 @@ class LocaleMiddleware(DjangoLocaleMiddleware):
         if not (i18n_patterns_used and language_from_path):
             patch_vary_headers(response, ('Accept-Language',))
         response.setdefault('Content-Language', language)
+
+        # An additional responsibility of our override of this middleware is to save the user's language
+        # preference in a cookie. That determination is made by code handling the request
+        # and saved in the _set_language flag so that it can be used here.
+        if hasattr(request, '_set_language'):
+            response.set_cookie(settings.LANGUAGE_COOKIE_NAME, request._set_language)
+
         return response
 
 class RateLimitMiddleware(MiddlewareMixin):